Split 'message' and 'constraint' specifications
